unavoidable-docs/requirements.txt
# Unavoidable Documentation System Dependencies

# Database
psycopg2-binary>=2.9.0

# File watching
watchdog>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# AST parsing and code analysis
astroid>=2.15.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging
colorlog>=6.7.0

# Optional: Web dashboard
flask>=2.3.0
flask-cors>=4.0.0

# Optional: Email reports
sendgrid>=6.10.0